sprint12 Доделаны правки по переиспользованию Adapter
