+++ app/src/main/res/layout/track.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              xmlns:app="http://schemas.android.com/apk/res-auto"
               xmlns:tools="http://schemas.android.com/tools"
               android:orientation="horizontal"
               android:layout_width="match_parent"
@@ -40,7 +39,7 @@
                     android:layout_weight="1"
                     android:ellipsize="end"
                     android:maxLines ="1"
-                    tools:text="Artist Name" />
+                    tools:text="Artist Name"/>
 
             <ImageView
                     style="@style/SecondaryImage.GreyWhite"
@@ -52,8 +51,7 @@
                     style="@style/SecondaryTextView.GreyWhite.Track"
                     android:layout_weight="0"
                     android:id="@+id/trackTime"
-                    tools:text="Time"
-            />
+                    tools:text="Time"/>
         </LinearLayout>
 
     </LinearLayout>
@@ -63,7 +61,6 @@
             android:src="@drawable/arrow_forward"
             android:layout_marginEnd="23dp"
             android:layout_weight="0"
-            android:layout_gravity="end|center_vertical"
-    />
+            android:layout_gravity="end|center_vertical"/>
 
 </LinearLayout>
